# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Eko CP",
        "description": "Dokumentasi API untuk data Fase, Mapel, dan CP.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server utama"
        }
    ],
    "paths": {
...
        "/api/home": {
            "get": {
                "tags": [
                    "Home"
                ],
                "summary": "Menampilkan data dashboard utama",
                "operationId": "d65ef76cbb32ae304c5a24c525b95adb",
                "parameters": [
                    {
                        "name": "Device-ID",
                        "in": "header",
                        "description": "ID perangkat pengguna (opsional)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "d6f12b6e-0ad3-4b54-90d2-90e4123c1ef2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dashboard data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "stats": {
                                                    "properties": {
                                                        "total_tema": {
                                                            "type": "integer",
                                                            "example": 12
                                                        },
                                                        "total_kuis": {
                                                            "type": "integer",
                                                            "example": 4
                                                        },
                                                        "total_video": {
                                                            "type": "integer",
                                                            "example": 9
                                                        },
                                                        "total_cp": {
                                                            "type": "integer",
                                                            "example": 6
                                                        },
                                                        "total_kaidah": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "total_refleksi": {
                                                            "type": "integer",
                                                            "example": 8
                                                        },
                                                        "random_video": {
                                                            "type": "object"
                                                        },
                                                        "device": {
                                                            "type": "object"
                                                        },
                                                        "kuis_selesai": {
                                                            "type": "integer",
                                                            "example": 2
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "featured_tema": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "active_kuis": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/search": {
            "get": {
                "tags": [
                    "Home"
                ],
                "summary": "Mencari data tema, kuis, dan video berdasarkan keyword",
                "operationId": "7e0945ccea0cb30ffa6973d2d4de2c1d",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Kata kunci pencarian",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "shalat"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "tema": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "kuis": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "video": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}